refactor file structure and folder names
