+++ lib/screens/pages/annotation.dart
@@ -1,9 +1,9 @@
 import 'dart:typed_data';
 import 'dart:ui' as ui;
-import 'package:binsight_ai/widgets/heading.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:binsight_ai/widgets/free_draw.dart';
+import 'package:binsight_ai/widgets/heading.dart';
 
 /// Page used for annotating an individual detection image
 class AnnotationPage extends StatefulWidget {
