+++ lib/screens/pages/detection_page.dart
@@ -1,9 +1,9 @@
-import 'package:binsight_ai/screens/main/annotation.dart';
-import 'package:binsight_ai/widgets/detections.dart';
 import 'package:flutter/material.dart';
-import 'package:binsight_ai/database/models/detection.dart';
-import 'package:binsight_ai/widgets/heading.dart';
 import 'package:go_router/go_router.dart';
+import 'package:binsight_ai/database/models/detection.dart';
+import 'package:binsight_ai/screens/pages/annotation.dart';
+import 'package:binsight_ai/widgets/detections.dart';
+import 'package:binsight_ai/widgets/heading.dart';
 
 /// Displays information about a single detection.
 class DetectionPage extends StatelessWidget {
