+++ lib/screens/pages/detections_page.dart
@@ -1,5 +1,5 @@
-import 'package:binsight_ai/database/models/detection.dart';
 import 'package:flutter/material.dart';
+import 'package:binsight_ai/database/models/detection.dart';
 import 'package:binsight_ai/widgets/detections.dart';
 import 'package:binsight_ai/widgets/heading.dart';
