+++ lib/screens/pages/stats_page.dart
