+++ lib/screens/splash/wifi_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:binsight_ai/widgets/wifi_configuration_widget.dart';
 import 'package:flutter_blue_plus/flutter_blue_plus.dart';
+import 'package:binsight_ai/widgets/wifi_configuration_widget.dart';
 
 /// Displays the WiFi configuration page with background and padding.
 class WifiPage extends StatelessWidget {
