+++ lib/widgets/wifi_configuration_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:binsight_ai/screens/bluetooth/bluetooth_page.dart';
+import 'package:binsight_ai/screens/pages/bluetooth_page.dart';
 import 'package:flutter_blue_plus/flutter_blue_plus.dart';
 import 'dart:convert';
